% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TweeterBiGram2Cloud.R
\name{TweeterBiGram2Cloud}
\alias{TweeterBiGram2Cloud}
\title{Create wordcloud by wordcloud2.js from Tweets by timeline}
\usage{
TweeterBiGram2Cloud(handle, n.tweets=5000, Bigrams = 500)
}
\arguments{
\item{handle}{A twitter handle with the timeline to examine}

\item{n.tweets}{The number of tweets to examine, default is 5000.}

\item{Bigrams}{The number of bigrams to plot, default is 500.}
}
\description{
Function for Creating wordcloud from a twitter handle through the Twitter API
}
\examples{
library(TweetClouds)
# Global variables can go here



TweeterBiGram2Cloud("realDonaldTrump")
TweeterBiGram2Cloud("realDonaldTrump", n.tweets=9000, Bigrams=350)

}
